requirements.txt
backpack-for-pytorch==1.3.0
gpustat==0.6.0
pandas==1.3.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.62.1